# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Parbati Enterprises Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin+)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserRecord"
                required: [data]
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin+)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/toggle-active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      summary: Toggle user active status (admin+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  isActive:
                    type: boolean
                required: [isActive]
...
  /masters/vendors:
    get:
      operationId: listVendors
      tags: [masters]
      summary: List vendors (searchable, paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Vendor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
    post:
      operationId: createVendor
      tags: [masters]
      summary: Create vendor (admin+)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /masters/vendors/{id}:
    get:
      operationId: getVendor
      tags: [masters]
      summary: Get vendor by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Vendor"
                required: [data]
    put:
      operationId: updateVendor
      tags: [masters]
      summary: Update vendor (admin+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorCreateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /masters/vendors/{id}/toggle-active:
    patch:
      operationId: toggleVendorActive
      tags: [masters]
      summary: Toggle vendor active status (admin+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  isActive:
                    type: boolean
                required: [isActive]
  /masters/vendors/{id}/vouchers:
    get:
      operationId: getVendorVouchers
      tags: [masters]
      summary: Get all vouchers for a vendor (accounts+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [payment, receive]
        - name: paymentStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Vouchers for vendor with vendor metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorVouchersResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    IdResponse:
      type: object
      properties:
        id:
          type: integer
      required: [id]
...
    UserRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [expense_entry, accounts, admin, superadmin]
        canViewHistory:
          type: integer
        isActive:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, canViewHistory, isActive]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [expense_entry, accounts, admin, superadmin]
        canViewHistory:
          type: boolean
      required: [name, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [expense_entry, accounts, admin, superadmin]
        canViewHistory:
          type: boolean
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isActive:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, isActive]
    VendorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    VendorCreateRequest:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
      required: [name]
...
    ExpenseListItem:
      type: object
      properties:
        expenseId:
          type: integer
        voucherType:
          type: string
          enum: [payment, receive]
        projectName:
          type: string
        vendorName:
          type: string
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        particularName:
          type: string
        uomName:
          type: string
        quantity:
          type: string
        pricePerUnit:
          type: string
        amount:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        paymentStatusName:
          type: string
        paymentStatusId:
          type: integer
        versionId:
          type: integer
        versionNo:
          type: integer
        approvedAt:
          type: string
          format: date-time
          nullable: true
        finalizedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - expenseId
        - voucherType
        - projectName
        - vendorName
        - voucherNumber
        - expenseDate
        - particularName
        - uomName
        - quantity
        - pricePerUnit
        - amount
        - paymentStatusName
        - paymentStatusId
        - versionId
        - versionNo
...
    VendorVouchersResponse:
      type: object
      properties:
        vendor:
          $ref: "#/components/schemas/Vendor"
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseListItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [vendor, data, total, page, limit]